Relocated RangeFilterCRUD, and it's test To allow default modifiers on the filter classes, everything must be in the same package
