+++ src/test/java/dk/sdu/se_f22/sortingmodule/range/rangepublic/crud/RangeFilterCreatorTest.java
@@ -1,6 +1,7 @@
 package dk.sdu.se_f22.sortingmodule.range.rangepublic.crud;
 
 import dk.sdu.se_f22.sortingmodule.range.exceptions.InvalidFilterException;
+import dk.sdu.se_f22.sortingmodule.range.rangepublic.RangeFilterCRUD;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.*;
