+++ src/test/java/dk/sdu/se_f22/sortingmodule/range/rangepublic/crud/RangeFilterDeleterTest.java
@@ -24,6 +24,10 @@ public class RangeFilterDeleterTest {
     int id2;
     int id3;
 
+
+
+
+
     @BeforeEach
     public void setup (){
         cleanUp();
