+++ src/test/resources/dk/sdu/se_f22/sortingmodule/range/rangepublic/DoubleFilter.csv
